requirements.txt
flask
mutagen
pillow
flask-login